backend/requirements.txt
flask
flask-cors
flask-jwt-extended
pymongo
python-dotenv
werkzeug
